Compiler warning

+++ Source/Geometry.m
@@ -129,7 +129,7 @@ return(theScaledImageRect);
 
 NSString *NSStringFromCIntegerPoint(CIntegerPoint inPoint)
 {
-return([NSString stringWithFormat:@"%d,%d", inPoint.x, inPoint.y]);
+return([NSString stringWithFormat:@"%ld,%ld", (long)inPoint.x, (long)inPoint.y]);
 }
 
 extern CIntegerPoint CIntegerPointFromString(NSString *inString)
